using uint16 for matsize

+++ ESLab_Assign01_Group11/code/dsp/helloDSP_config.h
@@ -54,7 +54,7 @@ typedef struct ControlMsg
     MSGQ_MsgHeader header;
     Uint16 command;
     Char arg1[ARG_SIZE];
-    Uint32 mat[ROW_SIZE][COL_SIZE];
+    Uint16 mat[ROW_SIZE][COL_SIZE];
 } ControlMsg;
 
 /* Messaging buffer used by the application.
